+++ accounting/RELEASES.md
@@ -308,3 +308,13 @@ Based on:
 - [typescript v0.8.0] accounting
 ### Releases
 - [NPM v0.8.0] https://www.npmjs.com/package/ryan-accounting/v/0.8.0 - accounting
+
+## 2025-02-13 01:39:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.490.0 (2.512.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.9.0] accounting
+### Releases
+- [NPM v0.9.0] https://www.npmjs.com/package/ryan-accounting/v/0.9.0 - accounting